Making Validators working

+++ programming1/5lab/src/data/classes/Validators/classes/CityValidator/AreaCityValidator.java
@@ -2,9 +2,9 @@ package src.data.classes.Validators.classes.CityValidator;
 
 import src.data.classes.Validators.interfaces.ValidatorInterface;
 
-public class AreaCityValidator implements ValidatorInterface<Long> {
+public class AreaCityValidator implements ValidatorInterface<Integer> {
     @Override
-    public boolean validate(Long variable) {
+    public boolean validate(Integer variable) {
         return (variable != null && variable > 0);
     }
