+++ programming1/5lab/src/data/classes/Validators/classes/CityValidator/TelephoneCodeCityValidator.java
@@ -2,9 +2,9 @@ package src.data.classes.Validators.classes.CityValidator;
 
 import src.data.classes.Validators.interfaces.ValidatorInterface;
 
-public class TelephoneCodeCityValidator implements ValidatorInterface<Integer> {
+public class TelephoneCodeCityValidator implements ValidatorInterface<Long> {
     @Override
-    public boolean validate(Integer variable) {
+    public boolean validate(Long variable) {
         return (variable > 0 && variable < 100000);
     }   
 }
